Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/moving-between-documents/before-prepare-iframe-parse-error-inline-module-expected.txt

CONSOLE MESSAGE: SyntaxError: Left side of assignment is not a reference.


PASS Sanity check around top-level Window
PASS Eval: Move parse-error inline module script to iframe before-prepare
PASS <script> load: Move parse-error inline module script to iframe before-prepare
FAIL <script> error: Move parse-error inline module script to iframe before-prepare assert_unreached: Script error evennt fired unexpectedly Reached unreachable code
PASS window error: Move parse-error inline module script to iframe before-prepare
